models/blueprint: Extract insertIntoFile() function into separate file
