+++ lib/models/blueprint.js
@@ -10,6 +10,7 @@ var MarkdownColor       = require('../utilities/markdown-color');
 var printableProperties = require('../utilities/printable-properties').blueprint;
 var sequence            = require('../utilities/sequence');
 var printCommand        = require('../utilities/print-command');
+var insertIntoFile      = require('../utilities/insert-into-file');
 var fs                  = require('fs-extra');
 var existsSync          = require('exists-sync');
 var inflector           = require('inflection');
@@ -1161,60 +1162,8 @@ var Blueprint = CoreObject.extend({
     @return {Promise}
   */
   insertIntoFile: function(pathRelativeToProjectRoot, contentsToInsert, providedOptions) {
-    var fullPath          = path.join(this.project.root, pathRelativeToProjectRoot);
-    var originalContents  = '';
-
-    if (existsSync(fullPath)) {
-      originalContents = fs.readFileSync(fullPath, { encoding: 'utf8' });
-    }
-
-    var contentsToWrite   = originalContents;
-
-    var options           = providedOptions || {};
-    var alreadyPresent    = originalContents.indexOf(contentsToInsert) > -1;
-    var insert            = !alreadyPresent;
-    var insertBehavior    = 'end';
-
-    if (options.before) { insertBehavior = 'before'; }
-    if (options.after)  { insertBehavior = 'after'; }
-
-    if (options.force) { insert = true; }
-
-    if (insert) {
-      if (insertBehavior === 'end') {
-        contentsToWrite += contentsToInsert;
-      } else {
-        var contentMarker      = options[insertBehavior];
-        var contentMarkerIndex = contentsToWrite.indexOf(contentMarker);
-
-        if (contentMarkerIndex !== -1) {
-          var insertIndex = contentMarkerIndex;
-          if (insertBehavior === 'after') { insertIndex += contentMarker.length; }
-
-          contentsToWrite = contentsToWrite.slice(0, insertIndex) +
-                            contentsToInsert + EOL +
-                            contentsToWrite.slice(insertIndex);
-        }
-      }
-    }
-
-    var returnValue = {
-      path: fullPath,
-      originalContents: originalContents,
-      contents: contentsToWrite,
-      inserted: false
-    };
-
-    if (contentsToWrite !== originalContents) {
-      returnValue.inserted = true;
-
-      return writeFile(fullPath, contentsToWrite)
-        .then(function() {
-          return returnValue;
-        });
-    } else {
-      return Promise.resolve(returnValue);
-    }
+    var fullPath = path.join(this.project.root, pathRelativeToProjectRoot);
+    return insertIntoFile(fullPath, contentsToInsert, providedOptions);
   },
 
   _printCommand: printCommand,
